Test: 이미지 업로드 컴포넌트 테스트 파일 분리
